// frontend/src/api/generated/models/riskConfigResponse.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * vibe-quant API
 * OpenAPI spec version: 0.1.0
 */
import type { RiskConfigResponsePortfolioLevel } from './riskConfigResponsePortfolioLevel';
import type { RiskConfigResponseStrategyLevel } from './riskConfigResponseStrategyLevel';

export interface RiskConfigResponse {
  id: number;
  name: string;
  strategy_level: RiskConfigResponseStrategyLevel;
  portfolio_level: RiskConfigResponsePortfolioLevel;
  created_at: string;
}